# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  meal.Meal:
    properties:
      calories:
        default: 0
        type: number
      created_at:
        default: "null"
        type: string
      id:
        type: string
      image:
        default: "null"
        type: string
      ingredients:
        items:
          $ref: '#/definitions/types.Ingredient'
        type: array
      name:
        type: string
      nutrients:
        items:
          $ref: '#/definitions/types.Nutrient'
        type: array
      updated_at:
        default: "null"
        type: string
      userid:
        type: string
    required:
    - name
    type: object
...
  plans.Plan:
    properties:
      dayofweek:
        type: string
      exercise:
        items:
          type: string
        type: array
      id:
        type: string
      typeofplan:
        type: string
      userid:
        type: string
    required:
    - dayofweek
    - typeofplan
    - userid
    type: object
...
  types.Ingredient:
    properties:
      ingredientid:
        type: string
      numofservings:
        type: number
      unit:
        type: string
      value:
        type: number
    required:
    - ingredientid
    - numofservings
    - unit
    - value
    type: object
  types.Nutrient:
    properties:
      category:
        type: string
      name:
        type: string
      type:
        type: string
      unit:
        type: string
      value:
        type: number
    required:
    - name
    - type
    - unit
    - value
    type: object
  user.CreateUserDto:
    properties:
      activityLevel:
        default: 0
        description: default:"0" is not working
        type: integer
      age:
        maximum: 120
        minimum: 1
        type: integer
      created_at:
        default: "null"
        type: string
      email:
        type: string
      gender:
        type: string
      height:
        default: 0
        description: default:"0" is not working
        type: number
      hip:
        default: 0
        description: default:"0" is not working
        type: number
      neck:
        default: 0
        description: default:"0" is not working
        type: number
      password:
        type: string
      username:
        maxLength: 20
        minLength: 3
        type: string
      waist:
        default: 0
        description: default:"0" is not working
        type: number
      weight:
        default: 0
        description: default:"0" is not working
        type: number
    required:
    - age
    - email
    - gender
    - password
    - username
    type: object
  user.UpadateUsernamePasswordDto:
    properties:
      newPassword:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - password
    type: object
  user.UpdateBodyDto:
    properties:
      activityLevel:
        type: integer
      age:
        type: integer
      gender:
        type: string
      height:
        type: number
      hip:
        type: number
      neck:
        type: number
      waist:
        type: number
      weight:
        type: number
    type: object
  user.User:
    properties:
      activitylevel:
        default: 0
        description: default:"0" is not working
        type: integer
      age:
        maximum: 120
        minimum: 1
        type: integer
      created_at:
        default: "null"
        type: string
      email:
        type: string
      gender:
        type: string
      height:
        default: 0
        description: default:"0" is not working
        type: number
      hip:
        default: 0
        description: default:"0" is not working
        type: number
      id:
        type: string
      neck:
        default: 0
        description: default:"0" is not working
        type: number
      password:
        type: string
      username:
        maxLength: 20
        minLength: 3
        type: string
      waist:
        default: 0
        description: default:"0" is not working
        type: number
      weight:
        default: 0
        description: default:"0" is not working
        type: number
    required:
    - age
    - email
    - gender
    - password
    - username
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for GymsBro API.
  title: GymsBro API
  version: "1.0"
paths:
...
  /meals/user/{userid}:
    get:
      consumes:
      - application/json
      description: Get meals by user
      parameters:
      - description: User ID
        in: path
        name: userid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/meal.Meal'
            type: array
        "400":
          description: Bad Request
          schema: {}
      summary: Get meals by user
      tags:
      - meals
...
  /plans/user/{user_id}:
    get:
      consumes:
      - application/json
      description: Get a plan by user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/plans.Plan'
            type: array
        "400":
          description: Bad Request
          schema: {}
      summary: Get a plan by user
      tags:
      - plans
  /plans/user/{user_id}/{day}:
    delete:
      consumes:
      - application/json
      description: Delete a plan by user and day
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Day of week
        in: path
        name: day
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema: {}
      summary: Delete a plan by user and day
      tags:
      - plans
    get:
      consumes:
      - application/json
      description: Get a plan by user and day
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Day of week
        in: path
        name: day
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/plans.Plan'
        "400":
          description: Bad Request
          schema: {}
      summary: Get a plan by user and day
      tags:
      - plans
  /users:
    get:
      consumes:
      - application/json
      description: Get all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.User'
            type: array
        "400":
          description: Bad Request
          schema: {}
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a user
      parameters:
      - description: Create User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema: {}
      summary: Create a user
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema: {}
      summary: Delete a user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema: {}
      summary: Get a user
      tags:
      - users
  /users/{id}/body:
    patch:
      consumes:
      - application/json
      description: Update a user body
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: UpdateBody User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UpdateBodyDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema: {}
      summary: Update a user body
      tags:
      - users
  /users/{id}/usepass:
    patch:
      consumes:
      - application/json
      description: Update a user username and password
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: UpdateUsernamePassword User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UpadateUsernamePasswordDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema: {}
      summary: Update a user username and password
      tags:
      - users
schemes:
- http
- https
swagger: "2.0"